refactor: main api, queries, interfaces and custom hooks
