+++ src/Shared/Interfaces/IAPIDATASearch.ts
@@ -1,8 +1,10 @@
-export interface ISearchCoinsApiNorm {
+export interface IAPIDATASearch {
   id: string;
   name: string;
   api_symbol: string;
   symbol: string;
+  errors: string;
+  currency: string;
   market_cap_rank: number;
   thumb: string;
   large: string;
