+++ src/Shared/Interfaces/IMAINDATAMarket.ts
@@ -1,0 +1,8 @@
+import { IAPIDATAMarket } from "./IAPIDATAMarket";
+
+export interface IMAINDATAMarket {
+  paging: number;
+  errors: string;
+  queries: string;
+  results: [] | IAPIDATAMarket[];
+}
