+++ src/Shared/Interfaces/IMAINDATASearch.ts
@@ -1,0 +1,8 @@
+import { IAPIDATASearch } from "./IAPIDATASearch";
+
+export interface IMAINDATASearch {
+  paging: number;
+  errors: string;
+  queries: string;
+  results: [] | IAPIDATASearch[];
+}
